automation_scripts/requirements.txt
bs4
requests
pandas
openpyxl
newspaper3k
bertopic
scikit-learn
